datapath/maps: move mapsweeper to dedicated package
We'd like to use pkg/datapath/maps for generated MapSpec definitions,
which causes a bunch of import cycles through packages in maps/*, so move
this out to its own package. Concerns are also better separated this way.
