--- pkg/datapath/maps/doc.go
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright Authors of Cilium
-
-// Package maps performs various lifecycle operations related to maps in the
-// datapath.
-package maps
