+++ pkg/datapath/mapsweeper/doc.go
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright Authors of Cilium
+
+// Package mapsweeper implements cleanup of stale and/or disabled BPF maps.
+package mapsweeper
