+++ pkg/datapath/mapsweeper/map.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package maps
+package mapsweeper
 
 import (
 	"log/slog"
