+++ pkg/datapath/mapsweeper/map_test.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package maps
+package mapsweeper
 
 import (
 	"path/filepath"
